+++ examples/Console/EMotion/Motion/Motion_1axis_move_with_breakpoint.py
@@ -39,30 +39,39 @@ async def main():
         ## Parameters setting
         port = 0
         axis = 0
-        relative_position = 1
+        rel_posi_mode = 1
+        stop_decel = 0
+
+        ## Polarity and enable parameters
+        active_low = 0
         active_high = 1
-        stop_deceleration = 0
+
+        ## Breakpoint parameters
         start_position = 100
         pulse_width = 100
         pulse_period = 100
         pulse_number = 100
 
+        ## Motion open
         err = await dev.Motion_open_async(port)
         print("open_async:", err)
- 
+
+        ## Motion open configuration file
         err = await dev.Motion_opencfgFile_async('3AxisStage_2P.ini')
         print("opencfgFile_async:", err)
 
+        ## Motion load configuration file
         err = await dev.Motion_loadCfgFile_async()
         print("loadCfgFile_async:", err)
- 
-        err = await dev.Motion_cfgBreakPoint_async(port, axis, relative_position, active_high, start_position, pulse_width, pulse_period, pulse_number)
+
+        ## Motion configure
+        err = await dev.Motion_cfgBreakPoint_async(port, axis, rel_posi_mode, active_high, start_position, pulse_width, pulse_period, pulse_number)
         print("cfgBreakPoint_async:", err)
 
         err = await dev.Motion_enableBreakPoint_async(port, axis, int(True))
         print("enableBreakPoint_async:", err)
 
-        err = dev.Motion_cfgAxisMove_async(port, axis, relative_position, 5000)
+        err = dev.Motion_cfgAxisMove_async(port, axis, rel_posi_mode, target_position = 5000)
         print("cfgAxisMove_async:", err)
 
         err = await dev.Motion_rstEncoderPosi_async(port, axis)
@@ -70,7 +79,8 @@ async def main():
 
         err = await dev.Motion_enableServoOn_async(port, axis, int(True))
         print("enableServoOn_async:", err)
-        
+
+        ## Motion start
         err = await dev.Motion_startSingleAxisMove_async(port, axis)
         print("startSingleAxisMove_async:", err)
 
@@ -78,13 +88,15 @@ async def main():
         while move_status == 0:
             move_status = await dev.Motion_getMoveStatus_async(port, axis)
             print("getMoveStatus:", move_status) 
-            
-        err = await dev.Motion_stop_async(port, axis, stop_deceleration)
+
+        ## Motion stop          
+        err = await dev.Motion_stop_async(port, axis, stop_decel)
         print("stop_async:", err) 
  
         err = await dev.Motion_enableServoOn_async(port, axis, int(False))
         print("enableServoOn_async:", err)
 
+        ## Motion close
         err = await dev.Motion_close_async(port)
         print("close_async:", err) 
          
@@ -98,6 +110,14 @@ async def main():
     dev.close()
  
     return
-
+    
+def main_for_spyder(*args):
+    if asyncio.get_event_loop().is_running():
+        return asyncio.create_task(main(*args)).result()
+    else:
+        return asyncio.run(main(*args))
+ 
 if __name__ == '__main__':
-    asyncio.run(main())
+    asyncio.run(main()) ## Use terminal
+    # await main() ## Use Jupyter or IPython(>=7.0)
+    # main_for_spyder() ## Use Spyder