+++ examples/Console/EMotion/Motion/Motion_1axis_move_with_inposition.py
@@ -40,25 +40,33 @@ async def main():
         port = 0
         axis = 0
         two_pulse_mode = 1
-        relative_position = 1
-        dir_cw = 0
+        rel_posi_mode = 1
+        stop_decel = 0
+        ## Axis and encoder parameters
+        axis_dir_cw = 0
+        encoder_dir_cw = 0
+        ## Polarity and enable parameters
         active_low = 0
         active_high = 1
-        stop_deceleration = 0 
-
+        forward_enable_true = 1
+        reverse_enable_true = 1
+        inposi_enable_false = 0
+  
+        ## Motion open
         err = await dev.Motion_open_async(port)
         print("open_async:", err)
 
-        err = await dev.Motion_cfgAxis_async(port, axis, two_pulse_mode, dir_cw, dir_cw, active_low)
+        ## Motion configure
+        err = await dev.Motion_cfgAxis_async(port, axis, two_pulse_mode, axis_dir_cw, encoder_dir_cw, active_low)
         print("cfgAxis_async:", err)
             
-        err = await dev.Motion_cfgLimit_async(port, axis, int(True), int(True), active_high)
+        err = await dev.Motion_cfgLimit_async(port, axis, forward_enable_true, reverse_enable_true, active_high)
         print("cfgLimit_async:", err)
 
-        err = await dev.Motion_cfgInposi_async(port, axis, int(False), active_low)
+        err = await dev.Motion_cfgInposi_async(port, axis, inposi_enable_false, active_low)
         print("cfgInposi_async:", err)
             
-        err = dev.Motion_cfgAxisMove_async(port, axis, relative_position, 5000)
+        err = dev.Motion_cfgAxisMove_async(port, axis, rel_posi_mode, target_position = 5000)
         print("cfgAxisMove_async:", err)
 
         err = await dev.Motion_rstEncoderPosi_async(port, axis)
@@ -67,6 +75,7 @@ async def main():
         err = await dev.Motion_enableServoOn_async(port, axis, int(True))
         print("enableServoOn_async:", err)
 
+        ## Motion start
         err = await dev.Motion_startSingleAxisMove_async(port, axis)
         print("startSingleAxisMove_async:", err)
 
@@ -74,13 +83,15 @@ async def main():
         while move_status == 0:
             move_status = await dev.Motion_getMoveStatus_async(port, axis)
             print("getMoveStatus:", move_status) 
-            
-        err = await dev.Motion_stop_async(port, axis, stop_deceleration)
+
+        ## Motion stop
+        err = await dev.Motion_stop_async(port, axis, stop_decel)
         print("stop_async:", err)
 
         err = await dev.Motion_enableServoOn_async(port, axis, int(False))
         print("enableServoOn_async:", err)
-        
+
+        ## Motion close
         err = await dev.Motion_close_async(port)
         print("close_async:", err) 
          
@@ -95,5 +106,13 @@ async def main():
  
     return
 
+def main_for_spyder(*args):
+    if asyncio.get_event_loop().is_running():
+        return asyncio.create_task(main(*args)).result()
+    else:
+        return asyncio.run(main(*args))
+ 
 if __name__ == '__main__':
-    asyncio.run(main())
+    asyncio.run(main()) ## Use terminal
+    # await main() ## Use Jupyter or IPython(>=7.0)
+    # main_for_spyder() ## Use Spyder